Fix a bug where complex key bindings are not working in tmux
By using xxd, we could know how complex key bindings are represented in
terminal.
For example, if we want to know byte streams of Ctrl + F5, we just run
xxd in bash, press <Ctrl + F5> and enter.  Then, 1b5b 3135 3b35 7e0a
will be printed. The last 0a is a return key, so remaining 7 bytes are
what we wanted. The only left thing is, google ascii table and
translate 7 bytes to '<esc>[15;5~'.