Add jq package to Darwin, Debian, and Redhat package managers
Add the jq package to the Darwin, Debian, and Redhat package managers. This package provides a command-line JSON processor which will be useful in data analysis. The commit also includes updates to the Debian package manager with the addition of the mediainfo package, which provides technical and tag information for video and audio files.