Create asset_strategy and encode.sh scripts
The commit introduces two new scripts: `asset_strategy` and `encode.sh`. The `asset_strategy` script is a Python script that imports various modules, including `datetime` and `subprocess`, and runs a Docker command using `subprocess.run`. The `encode.sh` script is a shell script that takes an input file path as an argument and encodes it using the `linuxserver/ffmpeg` Docker image. The script uses various options and flags to specify the input and output files, as well as the desired encoding format.

These scripts have been added to the repository in order to provide functionality related to asset strategy and video encoding.