Avoid yelling about pynvim
This commit adds logic to avoid getting warnings about pynvim when using neovim. The code checks if the `$VIRTUAL_ENV` variable exists and sets the `vim.g.python3_host_prog` accordingly. If `$VIRTUAL_ENV` exists, the code finds the path to python3 and sets it as the `vim.g.python3_host_prog`. If `$VIRTUAL_ENV` does not exist, the code simply sets the `vim.g.python3_host_prog` to the path of python3.