Add new device configuration in Karabiner
Configured a new device with address "d6-6f-1d-c9-65-2b" to be recognized as both a keyboard and pointing device in the `karabiner.json` file. Enabled horizontal and vertical mouse wheel flipping for this specific device, and set it not to be ignored. These changes ensure proper handling and customization of device input behavior through Karabiner.

BTW, the new device is a bluetooth-connected Tex Shinobi.