Fix LC_CTYPE setting for k9s compatibility
Previously, `LC_CTYPE` was set to `en_US.UTF-8` for the k9s ASCII diagram in the beginning of the script, but it was later overwritten by `ko_KR.UTF-8`. This change fixes the LC_CTYPE setting by placing `en_US.UTF-8` at the end of the locale configurations, thus ensuring that the correct locale is used for k9s ASCII diagrams. The redundant initial setting has been removed for clarity.