Add qpdf dependency and removePdfPassword script
Added 'qpdf' to the list of packages in both Darwin and Debian package managers to ensure it is installed as a dependency.
Created a new script 'removePdfPassword' for removing passwords from PDF files using qpdf.
The script prompts the user to enter a password and decrypts the specified PDF file, saving the decrypted version with a modified filename.
This addition allows for more efficient handling of secured PDF files.